CPU: Move PC and clock manipulation within methods
This moves PC and clock manipulation as far to the end of functions
as can easily be done.

It also makes the order uniform to be PC change, then clock change.

The first edit protects against premature PC increments, when calling
function which rely on the PC to be unchanged.

The second edit should not have any impact on functionality.